fix: Formalize snake-casing rules
The default rules are:

1. If a character is a spacing character, `-`, or `_`, place `_` into
   the output.
2. If a character is any other form of punctuation, remove it from the
   output.
3. Flip the character casing, if needed, then place it into the output.

`lsnake` and `usnake` now also take an additional parameter `extra`.
This field enables checking for additional characters to be cast into
`_` before placement into output. For example, when constructing the
include-guard:

    char *result = usnake(tmp, "./");

This will flag `.` and `/` as requiring replacement by `_` in the
output. To skip any such checks and use the defaults, simply pass
`NULL`:

    return usnake(opts->leader, NULL);