build: Remove `pedantic` and `no-keyword-macro` flags
`pedantic` displays warnings related to static lifetimes of macro
invocations that are functionally irrelevant, as they are used in
externally visible data sections.

`no-keyword-macro` is a holdover from Apple clang development when using
a custom `sizeof` macro as an override to the builtin `sizeof` operator.
That override has been removed, as well.